[2단계 - 상품 관리 기능] 아코(안석환) 미션 제출합니다 (#326)
* docs: 리팩토링 목록 추가

* refactor: 수정 요청 http 메소드 patch에서 put으로 수정

* refactor: 테스트 케이스를 성공과 실패로 나누어서 관리

* docs: 구현할 기능 목록 추가

* refactor: null과 blank 에러를 공통 에러로 추출

* feat: 사용자 이메일 검증 추가

* feat: 사용자 비밀번호 검증 추가

* style: productdao 패키지 분리

* docs: 구현할 기능 목록 수정

* feat: 사용자 전체 조회 기능 추가

* style: 패키지 구조 수정

* feat: 사용자 설정 페이지 연동

세부사항
- member 전체 조회 기능 화면과 연동

* style: product 패키지 분리

* refactor: product 필드 객체로 포장

* feat: cart 테이블 생성

* refactor: 공통된 예외처리 방식으로 수정

세부사항
- product 전용 예외처리에서 공통으로 사용할 수 있는 예외처리로 확장

* fix: 테스트가 실패하는 오류 수정

세부사항
- 결과를 비교하는데 있어서 하드 코딩 제거

* feat: db cart 추가 기능 구현

* feat: db cart member와 productId로 조회하는 기능 구현

* feat: db cart의 count를 update하는 기능 구현

* feat: db email로 member를 조회하는 기능 구현

* feat: MemberRequest의 ArgumentResolver 추가

* feat: 장바구니 상품 추가 기능 화면 연동

* feat: db 사용자 각자의 장바구니 조회 기능 구현

* feat: 사용자 각자의 장바구니 조회 기능 화면 연동

* feat: db 장바구니 상품 삭제 기능 구현

* feat: 장바구니 상품 삭제 기능 화면 연동

* test: 장바구니 페이지 접속 테스트

* test: 테스트 더미 데이터 추가

* style: cartRequest에서 cartItemRequest로 파일명 변경

* docs: 기본 데이터 추가

* test: 장바구니 인수 테스트 추가

* style : 불필요한 import문 삭제 및 가독성 향상

* style : 불필요한 로깅 삭제

* docs: 구현한 기능 목록 체크

* docs: 리팩토링할 목록 추가

* refactor: 인터셉터에서 인증 처리하기

* refactor: 사용자 조회 시 이메일과 비밀번호 모두 이용하기

* refactor: 상품의 crud 인가 추가

* refactor: basic 디코딩 하는 공통 로직 분리

* style: url 및 변수 명칭 명확하게 수정

* test: 상품 crud 권한 검증 테스트 및 상품 읽기 테스트 추가

* test: 테스트 검증 부분 상세화

* style : 불필요한 import문 삭제 및 가동성 향상

* style : 오타 수정

* refactor: member 요청 검증 추가

* refactor: 객체에게 메세지를 던지도록 구현

* refactor: 반환체를 제네릭에서 memberInfo로 변경

* style: 사용하지 않는 코드 삭제

* refactor: 중복되는 파일 하나로 통일

* test: 이메일 어노테이션 검증 테스트 추가

* refactor: 인터셉터가 memberDao가 아닌 memberService를 의존하도록 수정

* refactor: 메소드 명칭에 알맞은 파라미터를 가지게 수정

* docs: 구현한 기능 목록 체크

* refactor: 조회 테스트 검증시 문자열 검증 방식에서 dto를 이용해서 검증하는 방식으로 수정

* style: 가독성 향상을 위한 코드 정리

+++ src/main/resources/static/js/admin.js
@@ -46,10 +46,21 @@ form.addEventListener('submit', (event) => {
 
 // TODO: [1단계] 상품 관리 CRUD API에 맞게 변경
 const createProduct = (product) => {
+    const credentials = localStorage.getItem('credentials');
+    if (!credentials) {
+        alert('사용자 정보가 없습니다.');
+        window.location.href = '/settings';
+        return;
+    }
+
     axios.request({
         url: 'admin/products',
         method: 'post',
-        data: product
+        data: product,
+        headers: {
+            'Authorization': `Basic ${credentials}`,
+            'Content-Type': "application/json; charset=utf-8"
+        }
     }).then((response) => {
         window.location.reload();
     }).catch((error) => {
@@ -59,12 +70,22 @@ const createProduct = (product) => {
 
 // TODO: [1단계] 상품 관리 CRUD API에 맞게 변경
 const updateProduct = (product) => {
+    const credentials = localStorage.getItem('credentials');
+    if (!credentials) {
+        alert('사용자 정보가 없습니다.');
+        window.location.href = '/settings';
+        return;
+    }
     const { id } = product;
 
     axios.request({
+
         url: 'admin/products/' + id,
-        method: 'patch',
-        data: product
+        method: 'put',
+        data: product,
+        headers: {
+            'Authorization': `Basic ${credentials}`
+        }
     }).then((response) => {
         window.location.reload();
     }).catch((error) => {
@@ -74,9 +95,18 @@ const updateProduct = (product) => {
 
 // TODO: [1단계] 상품 관리 CRUD API에 맞게 변경
 const deleteProduct = (id) => {
+    const credentials = localStorage.getItem('credentials');
+    if (!credentials) {
+        alert('사용자 정보가 없습니다.');
+        window.location.href = '/settings';
+        return;
+    }
     axios.request({
         url: 'admin/products/' + id,
-        method: 'DELETE'
+        method: 'DELETE',
+        headers: {
+            'Authorization': `Basic ${credentials}`
+        }
     }).then((response) => {
         window.location.reload();
     }).catch((error) => {
